chore: fix search and contacts
