+++ examples/basic/vite.config.ts
@@ -1,12 +1,7 @@
 import { defineConfig } from "vite";
-import reactServerComponents from "fully-react/rsc-plugin";
+import react from "fully-react";
 import inspect from "vite-plugin-inspect";
 
 export default defineConfig((env) => ({
-	build: {
-		rollupOptions: {
-			input: { "entry-client": "src/entry-client.tsx" },
-		},
-	},
-	plugins: [inspect(), reactServerComponents()],
+	plugins: [inspect(), react()],
 }));
