+++ examples/client-only/vite.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from "vite";
-import react from "fully-react/rsc-plugin";
+import react from "fully-react";
 
 export default defineConfig({
 	plugins: [react()],
