+++ examples/impala/src/entry-client.tsx
@@ -1,4 +1,4 @@
-import { Root, mount } from "stream-react/web/entry";
+import { Root, mount } from "fully-react/web/entry";
 // Load the route modules as RSC and export for impala
 import("./components/button");
 mount(<Root />);
