+++ examples/mdx/vite.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from "vite";
-import react from "fully-react/rsc-plugin";
+import react from "fully-react";
 import mdx from "@cyco130/vite-plugin-mdx";
 export default defineConfig({
 	plugins: [mdx({}), react()],
