+++ examples/search/app/routes/index.tsx
@@ -1,7 +1,9 @@
-import type { PageProps } from "fully-react";
-import { component } from "stream-react/server";
+import type { PageProps } from "fully-react/app-router";
+import { component } from "fully-react/server";
 import { Suspense } from "react";
-import Search from "./Search";
+
+import Layout from "../layout";
+import Search from "../Search";
 
 export const SearchResults = component(async function SearchResults({
 	search,
@@ -12,14 +14,14 @@ export const SearchResults = component(async function SearchResults({
 	return <div>Search results for {search}</div>;
 });
 
-export function SearchPage({ searchParams }: PageProps) {
+export default function SearchPage({ searchParams }: PageProps) {
 	let search = searchParams["q"] ?? "";
 	return (
-		<>
+		<Layout title="Search">
 			<Search initialSearch={search} />
 			<Suspense fallback={<div>Loading...</div>}>
 				<SearchResults search={search} />
 			</Suspense>
-		</>
+		</Layout>
 	);
 }
