+++ examples/search/package.json
@@ -18,6 +18,7 @@
   "devDependencies": {
     "@types/react": "18.0.33",
     "@types/react-dom": "18.0.11",
+    "@vitejs/plugin-react": "^3.1.0",
     "typescript": "^5.0.3",
     "vite": "4.3.0"
   }
