+++ examples/server-counter/app/root.tsx
@@ -1,6 +1,6 @@
 import Counter from "./Counter";
 import { increment, getCount } from "./api";
-import { Assets } from "stream-react/assets";
+import { Assets } from "fully-react/assets";
 import "./root.css";
 
 export default async function Root() {
