+++ examples/server-counter/package.json
@@ -11,12 +11,12 @@
     "postinstall": "prisma generate"
   },
   "dependencies": {
+    "fully-react": "^0.0.2",
     "@prisma/client": "^4.12.0",
     "react": "18.3.0-next-b14f8da15-20230403",
     "react-dom": "18.3.0-next-b14f8da15-20230403",
     "react-error-boundary": "^4.0.3",
-    "react-server-dom-webpack": "18.3.0-next-b14f8da15-20230403",
-    "stream-react": "^0.0.1"
+    "react-server-dom-webpack": "18.3.0-next-b14f8da15-20230403"
   },
   "devDependencies": {
     "@types/react": "18.0.33",
