+++ examples/with-prisma-auth/app/root.tsx
@@ -3,10 +3,10 @@ import { sayHello } from "./api";
 import { getSession as getAuthSession } from "rsc-auth";
 import { SignInButton, SignOutButton } from "rsc-auth/components";
 import { authOptions } from "./auth";
-import { ErrorBoundary, ResetButton } from "stream-react/error-boundary";
-import { request } from "stream-react/request";
-import { component } from "stream-react/server";
-import { Assets } from "stream-react/assets";
+import { ErrorBoundary, ResetButton } from "fully-react/error-boundary";
+import { request } from "fully-react/request";
+import { component } from "fully-react/server";
+import { Assets } from "fully-react/assets";
 import Button from "./Button";
 import { cache } from "react";
