+++ test/template/app/root.tsx
@@ -1,4 +1,4 @@
-import { Assets } from "stream-react/assets";
+import { Assets } from "fully-react/assets";
 import "./root.css";
 
 export default function Root() {
